+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### Version 0.2.0 (4th April 2016)
+#### Added
+- Added possibility to set SDK prefix (for non-native purchase SDK usage only).
+- Added possibility to verify purchase by passing only transaction ID string.
+
+#### Changed
+- Improved logging by adding new log messages.
+
+---
+
 ### Version 0.1.0 (2nd March 2016)
 #### Added
 - Initial release of the adjust purchase SDK for iOS.